requirements.txt
azure-servicebus
azure-identity
python-dotenv
python-json-logger
aiohttp